refactor(notifier): extract message formatting into notifier_format.go
notifier.go was ~1.4k lines mixing several concerns. Move the cohesive
message-formatting block — messageContext, extractMessageContext/extractInt,
the messageFormatters table, all fmt* event formatters, and formatMessage —
verbatim into notifier_format.go (same package). Pure code move: no logic,
signature, or behavior change; notifier.go drops to ~1140 lines.

Test plan:
- go build ./... ; golangci-lint ./internal/notifier/... 0 issues
- go test ./internal/notifier/ pass
